+++ lib/src/extensions/flutter_extensions/flutter_extensions.dart
@@ -5,6 +5,7 @@ export 'focus_scope_extensions.dart';
 export 'list_widget.dart';
 export 'media_query_extension.dart';
 export 'navigation.dart';
+export 'on_numbers.dart';
 export 'padding.dart';
 export 'platform_extension.dart';
 export 'scaffold_messenger_extension.dart';