+++ lib/src/platform_env/platform_env.dart
@@ -1,0 +1,1 @@
+export 'platform_env_io.dart' if (dart.library.html) 'platform_env_web.dart';