+++ lib/src/src.dart
@@ -1,4 +1,6 @@
 export 'package:dart_helper_utils/dart_helper_utils.dart';
 
 export 'extensions/extensions.dart';
+export 'platform_env/platform_env.dart';
 export 'value_notifier/value_notifier.dart';
+export 'widgets/widgets.dart';