+++ lib/src/widgets/widgets.dart
@@ -1,0 +1,2 @@
+export 'gradient_widget.dart';
+export 'multi_tap_detector.dart';